NN-3644 fix bug and improve alert code
